Fix linters. Correct remove function error

+++ scripts/main.js
@@ -86,8 +86,10 @@ form.addEventListener('submit', (e) => {
 
 const list = document.getElementById('book-list');
 list.addEventListener('click', (e) => {
-  BookList.deleteBook(e.target);
-  BookStored.removeBook(e.target.parentNode.classList.value);
+  if (e.target.classList.contains('remove-btn')) {
+    BookList.deleteBook(e.target);
+    BookStored.removeBook(e.target.parentNode.classList.value);
+  }
 });
 
 const listPage = document.getElementById('list-link');
@@ -97,28 +99,31 @@ const allBookSection = document.getElementById('all-books');
 const contactSection = document.getElementById('contact-section');
 
 listPage.addEventListener('click', () => {
-  allBookSection.style.display = "flex";
-  form.style.display = "none";
-  contactSection.style.display = "none";
-  listPage.style.color = "blue";
-  addPage.style.color = "black";
-  contactPage.style.color = "black";
+  allBookSection.style.display = 'flex';
+  form.style.display = 'none';
+  contactSection.style.display = 'none';
+  listPage.style.color = 'blue';
+  addPage.style.color = 'black';
+  contactPage.style.color = 'black';
 });
 
 addPage.addEventListener('click', () => {
-  allBookSection.style.display = "none";
-  form.style.display = "flex";
-  contactSection.style.display = "none";
-  listPage.style.color = "black";
-  addPage.style.color = "blue";
-  contactPage.style.color = "black";
+  allBookSection.style.display = 'none';
+  form.style.display = 'flex';
+  contactSection.style.display = 'none';
+  listPage.style.color = 'black';
+  addPage.style.color = 'blue';
+  contactPage.style.color = 'black';
 });
 
 contactPage.addEventListener('click', () => {
-  allBookSection.style.display = "none";
-  form.style.display = "none";
-  contactSection.style.display = "flex";
-  listPage.style.color = "black";
-  addPage.style.color = "black";
-  contactPage.style.color = "blue";
+  allBookSection.style.display = 'none';
+  form.style.display = 'none';
+  contactSection.style.display = 'flex';
+  listPage.style.color = 'black';
+  addPage.style.color = 'black';
+  contactPage.style.color = 'blue';
 });
+
+const dt = new Date();
+document.getElementById('datetime').innerHTML = `${(`0${dt.getDate()}`).slice(-2)}.${(`0${dt.getMonth() + 1}`).slice(-2)}.${dt.getFullYear()} ${(`0${dt.getHours()}`).slice(-2)}:${(`0${dt.getMinutes()}`).slice(-2)}`;
